--- modules/consensus/grandpa/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "ismp-grandpa"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.2.2", default-features = false, features = [
-    "derive"
-] }
-primitive-types = { workspace = true }
-scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
-merkle-mountain-range = { package = "ckb-merkle-mountain-range", version = "0.5.2", default-features = false }
-finality-grandpa = { version = "0.16.0", features = ["derive-codec"], default-features = false }
-
-# polytope labs
-ismp = { git = "https://github.com/polytope-labs/ismp-rs", branch = "main", default-features = false }
-primitives = { package = "ismp-grandpa-primitives", path = "primitives", default-features = false }
-verifier = { package = "ismp-grandpa-verifier", path = "verifier", default-features = false}
-pallet-ismp = { path = "../pallet-ismp", default-features = false }
-
-
-# substrate
-frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-frame-system = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-trie = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-io = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-consensus-aura = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
-
-# cumulus
-cumulus-primitives-core = { git = "https://github.com/paritytech/cumulus", branch = "release-v0.9.420", default-features = false }
-
-ismp-primitives = { path = "../pallet-ismp/primitives", default-features = false }
-substrate-state-machine = { path = "../pallet-ismp/primitives/state-machine", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "codec/std",
-    "frame-support/std",
-    "frame-system/std",
-    "scale-info/std",
-    "cumulus-primitives-core/std",
-    "ismp/std",
-    "sp-trie/std",
-    "sp-consensus-aura/std",
-    "sp-runtime/std",
-    "sp-io/std",
-    "primitive-types/std",
-    "pallet-ismp/std",
-    "sp-core/std",
-    "primitives/std",
-    "verifier/std",
-    "merkle-mountain-range/std",
-    "ismp-primitives/std",
-    "substrate-state-machine/std",
-    "finality-grandpa/std",
-]
