+++ modules/consensus/grandpa/prover/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ismp-grandpa-prover"
+name = "grandpa-prover"
 version = "0.1.0"
 edition = "2021"
 authors = ["Polytope Labs <hello@polytope.technology>"]
@@ -10,22 +10,23 @@ publish = false
 [dependencies]
 hex = "0.4.3"
 anyhow.workspace = true
-serde = "1.0.144"
+serde = { workspace  = true, default-features = true}
 subxt = { workspace = true, default-features = true }
-codec = { package = "parity-scale-codec", version = "3.2.2", features = ["derive"] }
+codec = { workspace = true, default-features = true, features = ["derive"] }
 derive_more = "0.99.17"
 downcast-rs = "1.2.0"
-jsonrpsee = { version = "0.16.2", features = ["async-client", "jsonrpsee-ws-client"] }
-jsonrpsee-ws-client = "0.16.2"
+jsonrpsee = { version = "0.23", features = ["async-client", "jsonrpsee-ws-client"] }
+jsonrpsee-ws-client = "0.23"
 finality-grandpa = "0.16.0"
 
-sc-consensus-grandpa-rpc = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-sp-consensus-grandpa = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-sp-trie = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-sp-state-machine = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
+sc-consensus-grandpa-rpc = { workspace = true, default-features = true }
+sp-consensus-grandpa = { workspace = true, default-features = true }
+sp-runtime = { workspace = true, default-features = true }
+sp-core = { workspace = true, default-features = true }
+sp-trie = { workspace = true, default-features = true }
+sp-state-machine = { workspace = true, default-features = true }
 
 
-primitives = { package = "ismp-grandpa-primitives", path = "../primitives" }
-ismp = { git = "https://github.com/polytope-labs/ismp-rs", branch = "main" }
+grandpa-verifier-primitives = { workspace = true, default-features = true }
+ismp = { workspace = true, default-features = true }
+subxt-utils = { workspace = true, default-features = true }
