+++ modules/ismp/clients/grandpa/Cargo.toml
@@ -1,0 +1,59 @@
+[package]
+name = "ismp-grandpa"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+
+[dependencies]
+codec = { workspace = true, features = [
+    "derive"
+] }
+primitive-types = { workspace = true }
+scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
+merkle-mountain-range = { workspace = true }
+finality-grandpa = { version = "0.16.0", features = ["derive-codec"], default-features = false }
+
+# polytope labs
+ismp = { workspace = true }
+grandpa-verifier-primitives = { workspace = true  }
+grandpa-verifier = { workspace = true }
+pallet-ismp = { workspace = true }
+
+
+# substrate
+frame-support = { workspace = true }
+frame-system = { workspace = true }
+sp-trie = { workspace = true }
+sp-io = { workspace = true }
+sp-runtime = { workspace = true }
+sp-consensus-aura = { workspace = true }
+sp-core = { workspace = true }
+
+# cumulus
+cumulus-primitives-core = { workspace = true }
+
+substrate-state-machine = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+    "codec/std",
+    "frame-support/std",
+    "frame-system/std",
+    "scale-info/std",
+    "cumulus-primitives-core/std",
+    "ismp/std",
+    "sp-trie/std",
+    "sp-consensus-aura/std",
+    "sp-runtime/std",
+    "sp-io/std",
+    "primitive-types/std",
+    "pallet-ismp/std",
+    "sp-core/std",
+    "grandpa-verifier-primitives/std",
+    "grandpa-verifier/std",
+    "merkle-mountain-range/std",
+    "substrate-state-machine/std",
+    "finality-grandpa/std",
+]
